# swagger/openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "GauthierFitness \u2014 API Documentation",
        "version": "1.0.0",
        "description": "**API REST de GauthierFitness** \u2014 boutique e-commerce de produits fitness personnalisables (v\u00eatements, accessoires, nutrition).\n\n## Conventions\n- Authentification : **Laravel Sanctum** (`Bearer <token>`). Le token est obtenu via `POST /api/login` ou `POST /api/register`.\n- Format : JSON (`Accept: application/json`).\n- Codes d'erreur :\n  - `401` \u2014 non authentifi\u00e9\n  - `403` \u2014 interdit (ex : ressource d'un autre utilisateur)\n  - `404` \u2014 ressource introuvable\n  - `422` \u2014 donn\u00e9es de validation invalides\n- Rate limiting : `POST /api/contact` est limit\u00e9 \u00e0 5 requ\u00eates/minute/IP.\n\n## Domaines fonctionnels\n- **Auth** \u2014 inscription, connexion, d\u00e9connexion, v\u00e9rification.\n- **Catalogue** \u2014 produits publics (listing pagin\u00e9 + d\u00e9tail).\n- **Panier & Commande** \u2014 gestion du panier, paiement Stripe, suivi commande.\n- **Customisation** \u2014 sessions de personnalisation 2D/3D + assets (logos, images).\n- **IA** \u2014 g\u00e9n\u00e9ration d'images via OpenAI pour designs personnalis\u00e9s.\n- **Admin** \u2014 produits, commandes, stocks (r\u00e9serv\u00e9 aux administrateurs)."
    },
    "servers": [
        {
            "url": "http://localhost:8000/api"
        }
    ],
...
    "security": [
        {
            "Sanctum": []
        }
    ],
    "paths": {
...
        "/payment/intent": {
            "post": {
                "operationId": "stripe.createPaymentIntent",
                "summary": "Create a Stripe PaymentIntent for the current cart",
                "tags": [
                    "Paiement"
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "shipping": {
                                        "type": "object",
                                        "properties": {
                                            "firstname": {
                                                "type": "string"
                                            },
                                            "lastname": {
                                                "type": "string"
                                            },
                                            "address": {
                                                "type": "string"
                                            },
                                            "zip": {
                                                "type": "string"
                                            },
                                            "city": {
                                                "type": "string"
                                            },
                                            "country": {
                                                "type": "string"
                                            },
                                            "phone": {
                                                "type": [
                                                    "string",
                                                    "null"
                                                ]
                                            }
                                        },
                                        "required": [
                                            "firstname",
                                            "lastname",
                                            "address",
                                            "zip",
                                            "city",
                                            "country"
                                        ]
                                    }
                                },
                                "required": [
                                    "shipping"
                                ]
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "integer",
                                    "const": 200
                                }
                            }
                        }
                    },
                    "422": {
                        "$ref": "#/components/responses/ValidationException"
                    },
                    "401": {
                        "$ref": "#/components/responses/AuthenticationException"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "Sanctum": {
                "type": "http",
                "description": "Token Sanctum obtenu via `POST /api/login` ou `POST /api/register`. Pr\u00e9fix\u00e9 par `Bearer ` dans le header `Authorization`.",
                "scheme": "bearer"
            }
        },
...
        "responses": {
            "ValidationException": {
                "description": "Validation error",
                "content": {
                    "application/json": {
                        "schema": {
                            "type": "object",
                            "properties": {
                                "message": {
                                    "type": "string",
                                    "description": "Errors overview."
                                },
                                "errors": {
                                    "type": "object",
                                    "description": "A detailed description of each field that failed validation.",
                                    "additionalProperties": {
                                        "type": "array",
                                        "items": {
                                            "type": "string"
                                        }
                                    }
                                }
                            },
                            "required": [
                                "message",
                                "errors"
                            ]
                        }
                    }
                }
            },
...
            "AuthenticationException": {
                "description": "Unauthenticated",
                "content": {
                    "application/json": {
                        "schema": {
                            "type": "object",
                            "properties": {
                                "message": {
                                    "type": "string",
                                    "description": "Error overview."
                                }
                            },
                            "required": [
                                "message"
                            ]
                        }
                    }
                }
            },
...
        }
    }
}